+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add ESC + auto focus support for dialogs
 - Update Vue to 3.2.41
+- Add finder module (only for super admin users)
 
 ## 2.3.2 (2022-10-24)